Add files via upload
This model is created because our pneumonia detection model was able to identify pneumonia in chest x rays. But our model can't able to determine the input image is chest x ray or not. So we created this model with chest xray images,dog and cat images and trained the model to correctly identify the input image as chest x ray or not.